backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy[asyncio]
asyncpg
alembic
redis
celery[redis]
langchain
langchain-anthropic
langchain-google-genai
langchain-openai
anthropic
google-generativeai
openai
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
websockets
pillow
python-dotenv
flower
cryptography
email-validator
